refactor: move files related to Band
