+++ giraffe/src/components/Band/index.ts
@@ -1,0 +1,1 @@
+export {BandLayer} from './BandLayer'
